fix: don't allocate memory exponentially in col...
...lection mappings

`DefaultAtlasContext::extractCollectionMappings` has the a side effect
of modifying the mapping, thus for already expanded collection mappings
it expands again and replaces the mapping.

For example, given source document of:
```json
{ "array":
  [
    { "property": 1 },
    { "property": 2 },
    { "property": 3 }
  ]
}
```

And mapping containing a JSON field with path like `/array<>/property`.

First invocation modifies the mapping to contain three mappings:
 1. `/array<0>/property`
 2. `/array<2>/property`
 3. `/array<3>/property`

Since this is now replacing the previous mapping on the next invocation
this expands to:

 1. `/array<0>/property`
 2. `/array<1>/property`
 3. `/array<2>/property`
...
 9. `/array<8>/property`

This removes the mapping mutation side effect, though I'm not entirely
sure why it was there to begin with, could be for optimization?

Fixes #508