feat: Handle FieldGroup contributed via mapping definition
Fixes: #550

This allows transformations to change multiplicity, which is supposed to replace `COMBINE`/`SEPARATE` mode eventually. UI needs to add `Concatenate`/`Split` transformation instead of using `COMBINE`/`SEPARATE` mode, and for `Concatenate` case the source fields needs to be wrapped with `FieldGroup`, so that the `Concatenate` transformation is placed on a `FieldGroup` level.
[An example mapping is here](https://github.com/atlasmap/atlasmap/blob/master/runtime/itests/core/src/test/resources/mappings/atlasmapping-concatenate-split.xml)

This also allows transformations like `Add` mentioned in #461, which consumes multiple number fields and sum up. In combination with #338 we can also apply it for collection field.

At this point we don't support to have a collection field as a part of grouping, i.e. put collection field and non-collection field into a same group, so UI should keep prohibiting it. We can revisit it at some point.