fix: class cast when inspecting class with generics
For Kudu connector in Syndesis there are protobuf shaded classes
referenced from org.apache.kudu.client.KuduTable which use generics.

The code path leads to ParameterizedType describing a Comparable<T>
which cannot be cast to a Class.

This is a super simple workaround for this issue, not sure if it's 100%
proper.