feat: Support dynamic path expression resolution
Fixes: #1512
Instead of having expression as a transformation, this will bring `<Document ID>:</path/to/field>` syntax of path expression directly into expression parser. It allows expression to support more complex cases like FILTER and SELECT.
This partially addresses #1507, but in order to have it working for real usecase, #1562 has to be addressed for each FieldReader of java/json/xml.
Also, this supports specifying single index on a field (#1187).