chore: API documentation with OpenAPI 3.0
To get the API documentation published Maven plugins that support
OpenAPI 3.0 were introduced. The JAX-RS and swagger-core annotatations
are converted to OpenAPI 3.0 document based on the
`openapi-configuration.yaml` configuration file that lists the API
services. For the OpenAPI 3.0 document to Asciidoc document Maven plugin
from openapitools is used.

The end result is a slightly different Asciidoc document for the API,
and a slightly different HTML output.