feat(react-ui): add stories to atlasmap-core
There are two stories right now, aboout how to use Atlasmap in a
standalone way (it's code is directly used by the atlasmap-standalone
package) and in an embedded way (this code will be used by Syndesis as
is).

In general this should allow working on the core without relying on
the atlasmap-standalone package, easing the debugging.