feat: Camel Atlasmap component targeting Camel 3.1
This creates a new Camel component utilizing Camel 3.1, basically
copying the existing component and setting the version to Camel 3,
migrating where needed. Notable changes are that the in/out pattern on
Exchange is now deprecated and Exchange how has the `message` property.
With that the original message can be mutated with the new body
resulting from the transformation.

One other thing of note is that the Spring and Spring Boot version
levels are now bumped to versions used in Camel 3.1.