feat: Deliver DataSource metadata for runtime
It delivers DataSourceMetadata on each modules at runtime so it could be used to read/write a field. For example when it's writing XML target document, element order could be restricted by its schema.
AtlasRewritingXSVisitor works as a XSOM visitor and rewrites XML target document in valid order by following XML Schema. And it runs in post target process. It would be better to make XmlFieldWriter write a field in valid order in the first place, but it's not easy with current random writing fashion.
Fixes: #1709