fix(ui): make drag & drop quicker
The backend used by the DndProvider was the Touch one with support for
mouse events enabled. The Touch backend isn't the best when used on
"pure desktop" environments because of the need to understand if the
user is performing a click or a drag, which introduces some lag.

With this we perform some device capability detection relying on CSS
media query on the `hover` and `pointer` capabilities. This isn't
bulletproof but should make for a better experience than before.