fix: XML schema inspection type cache disables too many fields for FHIR
Fixes: #2406
Narrowed element/type caching to only apply vertically to avoid recursion. It causes OOM if it runs with the previous FHIR XSD, but it turned out the `xhtml:div` involved in FHIR XSD (!!!!!) is disabled in Syndesis FHIR connector. Did the same thing and now response time of FHIR inspection is realistic.