feat: PoC for Java DSL runtime support
Fixes: #2650
This introduces most basic support of AtlasMappingBuilder and AtlasField which allows user to write mapping logic in Java code. This changeset only enables simple read/write for non-collection field to bring the foundation in earlier. We can build a set of convenient API on top of this like an iterator for collection field, writing COMPLEX object in one invocation by comparing field name, apply transformations and etc.