fix: [BUG] Wrong index for preview target collection
Fixes: #3109
It turned out UI is sending unexpected field index when it's one-to-many with a collection target, which breaks the calculation of collection nest level at backend. Fixed to avoid adding an index if there's only one field at UI side, as well as a safe guard at backend side to ignore the index if there's only one field.