fix: support space in path to AtlasMap workspace when loading jars
fixes: #3324 ENTESB-15784

URL.getPath should not be used directly. it needs to go through toURi

see Javadoc https://docs.oracle.com/javase/7/docs/api/java/net/URL.html

The URL class does not itself encode or decode any URL components
according to the escaping mechanism defined in RFC2396. It is the
responsibility of the caller to encode any fields, which need to be
escaped prior to calling URL, and also to decode any escaped fields,
that are returned from URL. Furthermore, because URL has no knowledge of
URL escaping, it does not recognise equivalence between the encoded or
decoded form of the same URL

Note, the URI class does perform escaping of its component fields in
certain circumstances. The recommended way to manage the encoding and
decoding of URLs is to use URI, and to convert between these two classes
using toURI() and URI.toURL().

Signed-off-by: Aurélien Pupier <apupier@redhat.com>