fix: replace existing CVS inspection parameters
When changing the CVS parameters the document URI accumulates previously
set parameters. This leads to unwanted parameters remaining, or more
than one value for a parameter being used.

This makes sure that when setting the CVS inspection parameters the
existing parameters are first removed. This way only the desired new
parameters are used.