fix: notify of mapping changes in initialization
When initializing with a preloaded mapping JSON, if the mapping JSON
changed, say because the inspection parameters that have changed and
this is reflected in the document URI, we need to notify that the
mapping changed.
A scenario in which this issue is encountered is when CSV parameters
change and no other change is performed. The saved mapping in that case
contains the old URI in the CVS document.

Fixes #3755