fix: exponentially increasing validation requests
This adds two guards to the subscriptions via the `first` operator,
firstly to continue only after all the documents have been processed,
and secondly to get notifications only for the first time the
initialization was done. Otherwise the notification would be triggered
for each document times each time initialization was done, and with each
document another observer would be registered with the subject, all'
leading to an exponentially increasing number of requests to the server
to validate the mapping.

Fixes #3761