fix: AtlasMap is stuck after importing CSV on latest Google Chrome 105
Fixes: #4320

Google Chrome 105 made a breaking change and made some case fail which used to be working fine with older Chrome.
https://developer.chrome.com/articles/fetch-streaming-requests/

ky made a fix for that
https://github.com/sindresorhus/ky/issues/450

But since the newer ky no longer provides umd, we had to change the Jest configuration to deal with esm. Also we needed to introduce craco to add @babel/plugin-proposal-nullish-coalescing-operator which is required to update ky.